deploy: 배포

+++ src/main/java/devping/nnplanner/domain/monthmenu/service/MonthMenuService.java
@@ -172,7 +172,8 @@ public class MonthMenuService {
 
                     monthMenuHospitalRepository.delete(monthMenuHospital);
 
-                    if (!monthMenuHospital.getHospitalMenu().getCreatedBy().isEmpty()) {
+                    if (monthMenuHospital.getHospitalMenu().getCreatedBy() != null
+                        && !monthMenuHospital.getHospitalMenu().getCreatedBy().isEmpty()) {
                         hospitalMenuRepository.delete(monthMenuHospital.getHospitalMenu());
                     }
